LayoutTests/media/controls-drag-timebar-expected.txt

Test that dragging the timebar thumb causes seeks.

EVENT(canplaythrough)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
END OF TEST
